requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
faiss-cpu
pypdf
python-dotenv
streamlit
pandas
# fastembed